Added Turtle Stomping and Generic Enemy Script
The player can bounce off enemies on the "Enemy" layer.  The player has a switch statement to check the enemy's tag to know which script to search for.  Such as getting the Turtle script from enemies with the Turtle tag.

Enemies have a generic enemy script with functions for taking stomp and shot damage.  The individual enemy script, such as turtle, has their own stomp and shot damage functions that call the generic enemy ones.  In the case of the turtle, its shot function is currently blank as it can be destroyed by getting shot.  Enemies with a stomp or shot damage <= 0 run the Die method on their generic enemy script.  This method currently destroys the object.